Uglifier JS compressor blows up with missing ;
I was getting a JS error from this file but only when it's compressed.  I added in the semicolon and the error goes away.  The error was an unexpected var.  This might actually be an issue with uglifier, I'm not sure, but I figure even if it is, it's an easy enough fix here that I should just add it in.